Updated DITAP language with Traci's additions
The USDS and the Office of Federal Procurement Policy (OFPP) have partnered to develop a digital IT acquisition professional community (DITAP). The first component of this community is a training and certification program for contracting officers, which is currently in its second round of piloting this effort. The goal of this effort is that this development program will be offered in multiple agencies, by various training providers, or by the Federal Acquisition Institute. There is currently no open course at this time, but check back to see more information about upcoming offerings of this program.