Provide link to DITAP open source content
As requested by Traci, we'd like to place on TechFAR Hub the link to the open source DITAP content.  Not sure if it is best to put it within TechFAR Hub at the end of the paragraph about DITAP, or maybe swap it out where there is currently a link to a medium blog that Traci wrote about TechFAR Hub.  Maybe you can place it where ever it makes most sense.